requirements.txt
pymodbus==3.1.1
flask==2.3.0
werkzeug==2.3.0
RPi.GPIO==0.7.0adafruit-circuitpython-dht==4.0.2